requirements.txt
streamlit>=1.28.0
pytesseract>=0.3.10
Pillow>=10.3.0
opencv-python>=4.9.0
numpy>=1.26.0
pandas>=2.1.0
sqlalchemy>=2.0.0
transformers>=4.35.0
torch>=2.1.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
plotly>=5.18.0

